backend/requirements.txt
aiomysql==0.2.0
aioodbc==0.5.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asttokens==2.4.1
bcrypt==4.2.1
cached-property==1.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crypto==1.4.1
cryptography==43.0.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exchangelib==5.4.3
executing==2.1.0
fastapi==0.114.1
fastapi-cli==0.0.6
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
icecream==2.1.3
idna==3.8
iniconfig==2.0.0
isodate==0.7.2
Jinja2==3.1.4
ldap3==2.9.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mysql-connector-python==9.1.0
Naked==0.1.32
oauthlib==3.2.2
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyad==0.6.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyMySQL==1.1.1
pyodbc==5.1.0
pyspnego==0.11.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
python-tds==1.15.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
setuptools==75.2.0
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
sspilib==0.2.0
starlette==0.38.5
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.1
watchfiles==1.0.0
websockets==14.1
